+++ consensus/helpers.go
@@ -6,7 +6,6 @@ import (
 	"log"
 
 	"github.com/pokt-network/pocket/shared/codec"
-	"github.com/pokt-network/pocket/shared/debug"
 
 	"google.golang.org/protobuf/proto"
 
@@ -28,8 +27,8 @@ const (
 
 	ByzantineThreshold = float64(2) / float64(3)
 
-	HotstuffMessage = "consensus.HotstuffMessage"
-	UtilityMessage  = "consensus.UtilityMessage"
+	HotstuffMessageContentType = "consensus.HotstuffMessage"
+	UtilityMessageContentType  = "consensus.UtilityMessage"
 )
 
 var (
@@ -126,7 +125,6 @@ func (m *consensusModule) isOptimisticThresholdMet(n int) error {
 func (m *consensusModule) resetForNewHeight() {
 	m.Round = 0
 	m.Block = nil
-	m.TxResults = nil
 	m.highPrepareQC = nil
 	m.lockedQC = nil
 }
@@ -154,7 +152,7 @@ func (m *consensusModule) sendToNode(msg *typesCons.HotstuffMessage) {
 		m.nodeLogError(typesCons.ErrCreateConsensusMessage.Error(), err)
 		return
 	}
-	if err := m.GetBus().GetP2PModule().Send(cryptoPocket.AddressFromString(m.idToValAddrMap[*m.LeaderId]), anyConsensusMessage, debug.PocketTopic_CONSENSUS_MESSAGE_TOPIC); err != nil {
+	if err := m.GetBus().GetP2PModule().Send(cryptoPocket.AddressFromString(m.idToValAddrMap[*m.LeaderId]), anyConsensusMessage); err != nil {
 		m.nodeLogError(typesCons.ErrSendMessage.Error(), err)
 		return
 	}
@@ -167,7 +165,7 @@ func (m *consensusModule) broadcastToNodes(msg *typesCons.HotstuffMessage) {
 		m.nodeLogError(typesCons.ErrCreateConsensusMessage.Error(), err)
 		return
 	}
-	if err := m.GetBus().GetP2PModule().Broadcast(anyConsensusMessage, debug.PocketTopic_CONSENSUS_MESSAGE_TOPIC); err != nil {
+	if err := m.GetBus().GetP2PModule().Broadcast(anyConsensusMessage); err != nil {
 		m.nodeLogError(typesCons.ErrBroadcastMessage.Error(), err)
 		return
 	}