+++ p2p/module.go
@@ -8,7 +8,7 @@ import (
 	"github.com/pokt-network/pocket/p2p/stdnetwork"
 	typesP2P "github.com/pokt-network/pocket/p2p/types"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
-	"github.com/pokt-network/pocket/shared/debug"
+	"github.com/pokt-network/pocket/shared/messaging"
 	"github.com/pokt-network/pocket/shared/modules"
 	"github.com/pokt-network/pocket/telemetry"
 	"google.golang.org/protobuf/proto"
@@ -18,7 +18,7 @@ import (
 var _ modules.P2PModule = &p2pModule{}
 
 const (
-	P2PModuleName = "p2p"
+	p2pModuleName = "p2p"
 )
 
 type p2pModule struct {
@@ -82,7 +82,7 @@ func (m *p2pModule) GetBus() modules.Bus {
 }
 
 func (m *p2pModule) GetModuleName() string {
-	return P2PModuleName
+	return p2pModuleName
 }
 
 func (m *p2pModule) Start() error {
@@ -135,12 +135,11 @@ func (m *p2pModule) Stop() error {
 	return nil
 }
 
-func (m *p2pModule) Broadcast(msg *anypb.Any, topic debug.PocketTopic) error {
-	c := &debug.PocketEvent{
-		Topic: topic,
-		Data:  msg,
+func (m *p2pModule) Broadcast(msg *anypb.Any) error {
+	c := &messaging.PocketEnvelope{
+		Content: msg,
 	}
-	data, err := proto.Marshal(c)
+	data, err := proto.MarshalOptions{Deterministic: true}.Marshal(c)
 	if err != nil {
 		return err
 	}
@@ -149,12 +148,11 @@ func (m *p2pModule) Broadcast(msg *anypb.Any, topic debug.PocketTopic) error {
 	return m.network.NetworkBroadcast(data)
 }
 
-func (m *p2pModule) Send(addr cryptoPocket.Address, msg *anypb.Any, topic debug.PocketTopic) error {
-	c := &debug.PocketEvent{
-		Topic: topic,
-		Data:  msg,
+func (m *p2pModule) Send(addr cryptoPocket.Address, msg *anypb.Any) error {
+	c := &messaging.PocketEnvelope{
+		Content: msg,
 	}
-	data, err := proto.Marshal(c)
+	data, err := proto.MarshalOptions{Deterministic: true}.Marshal(c)
 	if err != nil {
 		return err
 	}
@@ -163,6 +161,7 @@ func (m *p2pModule) Send(addr cryptoPocket.Address, msg *anypb.Any, topic debug.
 }
 
 func (*p2pModule) ValidateConfig(cfg modules.Config) error {
+	// TODO (#334): implement this
 	return nil
 }
 
@@ -180,15 +179,14 @@ func (m *p2pModule) handleNetworkMessage(networkMsgData []byte) {
 		return
 	}
 
-	networkMessage := debug.PocketEvent{}
+	networkMessage := messaging.PocketEnvelope{}
 	if err := proto.Unmarshal(appMsgData, &networkMessage); err != nil {
 		log.Println("Error decoding network message: ", err)
 		return
 	}
 
-	event := debug.PocketEvent{
-		Topic: networkMessage.Topic,
-		Data:  networkMessage.Data,
+	event := messaging.PocketEnvelope{
+		Content: networkMessage.Content,
 	}
 
 	m.GetBus().PublishEventToBus(&event)