+++ shared/bus.go
@@ -3,7 +3,7 @@ package shared
 import (
 	"log"
 
-	"github.com/pokt-network/pocket/shared/debug"
+	"github.com/pokt-network/pocket/shared/messaging"
 	"github.com/pokt-network/pocket/shared/modules"
 )
 
@@ -22,6 +22,7 @@ type bus struct {
 	consensus   modules.ConsensusModule
 	telemetry   modules.TelemetryModule
 	logger      modules.LoggerModule
+	rpc         modules.RPCModule
 
 	runtimeMgr modules.RuntimeMgr
 }
@@ -38,6 +39,7 @@ func CreateBus(
 	consensus modules.ConsensusModule,
 	telemetry modules.TelemetryModule,
 	logger modules.LoggerModule,
+	rpc modules.RPCModule,
 ) (modules.Bus, error) {
 	bus := &bus{
 		channel: make(modules.EventsChannel, DefaultPocketBusBufferSize),
@@ -50,6 +52,7 @@ func CreateBus(
 		consensus:   consensus,
 		telemetry:   telemetry,
 		logger:      logger,
+		rpc:         rpc,
 	}
 
 	modules := map[string]modules.Module{
@@ -59,6 +62,7 @@ func CreateBus(
 		"utility":     utility,
 		"telemetry":   telemetry,
 		"logger":      logger,
+		"rpc":         rpc,
 	}
 
 	// checks if modules are not nil and sets their bus to this bus instance.
@@ -91,6 +95,7 @@ func CreateBusWithOptionalModules(
 	consensus modules.ConsensusModule,
 	telemetry modules.TelemetryModule,
 	logger modules.LoggerModule,
+	rpc modules.RPCModule,
 ) modules.Bus {
 	bus := &bus{
 		channel: make(modules.EventsChannel, DefaultPocketBusBufferSize),
@@ -103,6 +108,7 @@ func CreateBusWithOptionalModules(
 		consensus:   consensus,
 		telemetry:   telemetry,
 		logger:      logger,
+		rpc:         rpc,
 	}
 
 	maybeSetModuleBus := func(mod modules.Module) {
@@ -117,15 +123,16 @@ func CreateBusWithOptionalModules(
 	maybeSetModuleBus(consensus)
 	maybeSetModuleBus(telemetry)
 	maybeSetModuleBus(logger)
+	maybeSetModuleBus(rpc)
 
 	return bus
 }
 
-func (m bus) PublishEventToBus(e *debug.PocketEvent) {
+func (m bus) PublishEventToBus(e *messaging.PocketEnvelope) {
 	m.channel <- *e
 }
 
-func (m bus) GetBusEvent() *debug.PocketEvent {
+func (m bus) GetBusEvent() *messaging.PocketEnvelope {
 	e := <-m.channel
 	return &e
 }
@@ -156,6 +163,10 @@ func (m bus) GetTelemetryModule() modules.TelemetryModule {
 
 func (m bus) GetLoggerModule() modules.LoggerModule {
 	return m.logger
+}
+
+func (m bus) GetRPCModule() modules.RPCModule {
+	return m.rpc
 }
 
 func (m *bus) GetRuntimeMgr() modules.RuntimeMgr {