+++ shared/modules/types.go
@@ -22,6 +22,7 @@ type Config interface {
 	GetP2PConfig() P2PConfig
 	GetTelemetryConfig() TelemetryConfig
 	GetLoggerConfig() LoggerConfig
+	GetRPCConfig() RPCConfig
 }
 
 type BaseConfig interface {
@@ -77,6 +78,7 @@ type RPCConfig interface {
 	GetEnabled() bool
 	GetPort() string
 	GetTimeout() uint64
+	GetUseCors() bool
 }
 
 type PersistenceGenesisState interface {