+++ persistence/CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this module will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+TODO: consolidate `persistence/docs/CHANGELOG` and `persistence/CHANGELOG.md`
+
 ## [Unreleased]
+
+## [0.0.0.7] - 2022-11-01
+
+- Ported over storing blocks and block components to the Persistence module from Consensus and Utility modules
+- Encapsulated `TxIndexer` logic to the persistence context only
 
 ## [0.0.0.6] - 2022-10-06